Created the summarizer
